requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
grad-cam==1.5.5
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.1
pandas==3.0.2
pillow==12.1.1
pydantic==2.13.2
pydantic_core==2.46.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==70.2.0
simsimd==6.5.16
six==1.17.0
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
ttach==0.0.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
